requirements.txt
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dnspython==2.4.0
falcon==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
install==1.3.5
mailjet-rest==1.3.4
motor==3.2.0
msgspec==0.17.0
PyJWT==2.7.0
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.3
uvicorn==0.23.0
